[1.2.0] Project directory restructuring and guide track metric updates
Restructured the project directories, placing the main NetLogo file in the
repository's root and moving data and Python script files into the data/ and
scripts/ directories respectively. Also updated the script generating track
statistics to actually compute track reach the way it was previously described,
and added a slider to the NetLogo UI to set a maximum reach threshold, in
addition to the minimum one.

Signed-off-by: Helio Perroni Filho <helio@yorku.ca>