[5.3.0] Functional MRFC implementation
Completed a first iteration of the MRFC implementation. Not all the method's
parameters are included in the NetLogo model, and it's clear that it's
necessary to add a distinction between collisions and interventions (when the
convoy stops for safety reasons and has to be reconfigured), but it's a running
implementation none the less.

Signed-off-by: Helio Perroni Filho <helio@yorku.ca>